requirements.txt
pandas
matplotlib
seaborn
numpy
nltk
sklearn
collections
re
sklearn
string
keras
tqdm
pyspellchecker
gensim
